refactor: modularize constants and update import paths
